SDR-113 refactor: :recycle: components 폴더 구조 정리
