--- fe/src/components/BoxContainer/BoxContainer.styled.ts
@@ -1,6 +1,0 @@
-import styled from 'styled-components';
-
-export const Wrap = styled.div`
-  width: 100%;
-  padding: 12px;
-`;
