--- fe/src/components/BoxContainer/BoxContainer.tsx
@@ -1,12 +1,0 @@
-import { ReactNode } from 'react';
-import { Wrap } from './BoxContainer.styled';
-
-function BoxContainer({ children }: BoxContainerProps) {
-  return <Wrap>{children}</Wrap>;
-}
-
-export default BoxContainer;
-
-interface BoxContainerProps {
-  children: ReactNode;
-}
