--- fe/src/components/ImgUpload/Button/Button.tsx
@@ -1,7 +1,0 @@
-import { ButtonWrapper } from './Button.styled';
-
-function Button() {
-  return <ButtonWrapper>사진 업로드</ButtonWrapper>;
-}
-
-export default Button;
