--- fe/src/components/ImgUpload/ImgUpload.styled.ts
@@ -1,10 +1,0 @@
-import styled from 'styled-components';
-import { flexLayoutMixin } from 'utils/utils';
-
-export const ImgUploadWrapper = styled.div`
-  ${() => flexLayoutMixin('column', 'center', 'center')};
-  * {
-    margin-top: 10px;
-  }
-  margin: auto auto;
-`;
