+++ fe/src/components/ReviewDetail/Carousel/Carousel.styled.ts
