+++ fe/src/components/ReviewDetail/TotalRating/TotalRating.styled.ts
