+++ fe/src/components/ReviewDetail/TotalRating/TotalRating.tsx
