+++ fe/src/components/ReviewWrite/ImgUpload/ImgUpload.styled.ts
@@ -1,4 +1,13 @@
 import styled from 'styled-components';
+import { flexLayoutMixin } from 'utils/utils';
+
+export const ImgUploadWrapper = styled.div`
+  ${() => flexLayoutMixin('column', 'center', 'center')};
+  * {
+    margin-top: 10px;
+  }
+  margin: auto auto;
+`;
 
 export const ButtonWrapper = styled.button`
   background-color: #7d91d9;
