+++ fe/src/components/ReviewWrite/ImgUpload/ImgUpload.tsx
@@ -1,13 +1,12 @@
 import Icon from 'common/Icon';
-import Button from './Button/Button';
-import { ImgUploadWrapper } from './ImgUpload.styled';
+import { ImgUploadWrapper, ButtonWrapper } from './ImgUpload.styled';
 
 function ImgUpload() {
   return (
     <ImgUploadWrapper>
       <Icon icon="Photo" fill="black" />
       <div>사진을 업로드해주세요</div>
-      <Button />
+      <ButtonWrapper>사진 업로드</ButtonWrapper>;
     </ImgUploadWrapper>
   );
 }
