+++ fe/src/components/ReviewWrite/PostScope/PostScope.styled.ts
