+++ fe/src/components/ReviewWrite/PostScope/PostScope.tsx
